Move ScanTypes to own module
