+++ runtime/src/bank/tests.rs
@@ -40,7 +40,8 @@ use {
     solana_account_info::MAX_PERMITTED_DATA_INCREASE,
     solana_accounts_db::{
         accounts::AccountAddressFilter,
-        accounts_index::{AccountIndex, AccountSecondaryIndexes, IndexKey, ScanConfig, ScanError},
+        accounts_index::{AccountIndex, AccountSecondaryIndexes, IndexKey},
+        accounts_scan::{ScanConfig, ScanError},
         ancestors::Ancestors,
     },
     solana_client_traits::SyncClient,
